backend/requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.6

# Database & Cache
supabase==2.18.1
redis==6.4.0
asyncpg==0.29.0
sqlalchemy==2.0.43

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Validation & Serialization
pydantic==2.11.7
pydantic-settings==2.1.0
email-validator==2.1.0

# Cloud Storage
boto3==1.34.20  # For Cloudflare R2 (S3 compatible)

# Email
python-dotenv==1.1.1
aiosmtplib==3.0.1
jinja2==3.1.3

# Utilities
python-dateutil==2.9.0
pytz==2023.3
httpx==0.28.1

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
pylint==3.0.3

# Monitoring & Logging
python-json-logger==2.0.7

# PDF Generation
reportlab==4.0.8
pypdf==3.17.4

# Image Processing
pillow==10.2.0

# CORS
fastapi-cors==0.0.6